Fix clashing of user vs juju/system action keys.
When charms set action results/keys via action-set, it is possible that
the charm sets "special" action keys (i.e. stdout, stderr) via the
key=val args.  This caused a race condition between juju collecting
stdout/stderr from running the action hook (i.e.  worker/uniter/runner
runner.updateActionResults) and the hook's call to "action-set [key=val]..."
resulting in one of the action-set key+val or the hook stdout
winning and clobbering the other in the controller.  It also caused
other unexpected behavior if the charm set the stdout result key to the
empty string (just ommitting "stdout" key from results entirely) or
showing the content of the "stdout" result key under the capitalized
"Stdout" in the final client results printout.

To resolve all these issues we now check to ensure that none of the keys
being set via action-set are in a reserved set that juju uses for
special purposes under the hood and error out accordingly.

This was originally reported and investigated via
https://github.com/canonical/operator/issues/689.